third_party/blink/web_tests/http/tests/cache/svg-image-usecounters-in-cached-image.html

<!DOCTYPE html>
<script src="../../js-test-resources/js-test.js"></script>
<script>
// From third_party/blink/public/mojom/use_counter/metrics/web_feature.mojom
var SVGSMILAnimationInImageRegardlessOfCache = 768;
var SVGImage = 5492;

window.jsTestIsAsync = true;

window.onload = function() {
    var img = new Image();
    if (location.hash != '#loadWithCachedImage') {
        img.onload = function() {
            location.hash = 'loadWithCachedImage';
            window.location.reload();
        }
    } else {
        shouldBeFalse("internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache)");
        shouldBeFalse("internals.isUseCounted(document, SVGImage)");
        img.onload = function() {
            shouldBeTrue("internals.isUseCounted(document, SVGSMILAnimationInImageRegardlessOfCache)");
            shouldBeTrue("internals.isUseCounted(document, SVGImage)");
            finishJSTest();
        }
    }
    img.src = "resources/svg-with-animation.svg";
}
</script>
</html>
